Added app "Life in Weeks on One Sheet"
It shows the 80 years of a person's life. Each year is represented by
one line/row, and each week is represented by an X. The years the person
has already lived are crossed-off/struck-through.

The product is presented as a downloadable PDF so that one can print it
and keep for reminding themselves and marking remaining weeks off
themselves.

The webpage uses responsive design by using appropriate CSS breakpoints
for different screen sizes.

By default, it shows Apple Inc. cofounder Steve Jobs' number of years
marked off on the sheet.